# api/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API to manage events.",
        "title": "middleware/example",
        "contact": {
            "name": "Justine Bachelard.",
            "email": "justine.bachelard@ext.uca.fr"
        },
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
        "/events": {
            "get": {
                "description": "Get all events.",
                "tags": [
                    "events"
                ],
                "summary": "Get all events.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Event"
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "description": "Get an event.",
                "tags": [
                    "events"
                ],
                "summary": "Get an event.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event UUID formatted ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "422": {
                        "description": "Cannot parse id"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Event": {
            "type": "object",
            "properties": {
                "agenda_id": {
                    "description": "ID de l'agenda source",
                    "type": "string"
                },
                "description": {
                    "description": "Description complète",
                    "type": "string"
                },
                "end_date": {
                    "description": "Date/heure fin (DTEND)",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_modified": {
                    "description": "Dernière modification",
                    "type": "string"
                },
                "location": {
                    "description": "Salle (ex: \"IS_A104\")",
                    "type": "string"
                },
                "start_date": {
                    "description": "Date/heure début (DTSTART)",
                    "type": "string"
                },
                "summary": {
                    "description": "Titre du cours (ex: \"CM Big Data\")",
                    "type": "string"
                },
                "uid": {
                    "description": "UID de l'événement iCal",
                    "type": "string"
                }
            }
        }
    }
}